@@ -17,18 +17,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.3.12",
+    "@angular/animations": "14.3.0",
     "@angular/cdk": "13.3.9",
-    "@angular/common": "13.3.12",
-    "@angular/compiler": "13.3.12",
-    "@angular/core": "13.3.12",
+    "@angular/common": "14.3.0",
+    "@angular/compiler": "14.3.0",
+    "@angular/core": "14.3.0",
     "@angular/flex-layout": "6.0.0-beta.16",
-    "@angular/forms": "13.3.12",
+    "@angular/forms": "14.3.0",
     "@angular/http": "6.0.3",
     "@angular/material": "13.3.9",
-    "@angular/platform-browser": "13.3.12",
-    "@angular/platform-browser-dynamic": "13.3.12",
-    "@angular/router": "13.3.12",
+    "@angular/platform-browser": "14.3.0",
+    "@angular/platform-browser-dynamic": "14.3.0",
+    "@angular/router": "14.3.0",
     "angular-highcharts": "6.2.6",
     "angular-pipes": "8.0.0",
     "angular-tree-component": "7.2.0",
@@ -65,10 +65,10 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "13.3.11",
-    "@angular/cli": "^13.3.11",
-    "@angular/compiler-cli": "13.3.12",
-    "@angular/language-service": "13.3.12",
+    "@angular-devkit/build-angular": "14.2.11",
+    "@angular/cli": "^14.2.11",
+    "@angular/compiler-cli": "14.3.0",
+    "@angular/language-service": "14.3.0",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~10.3.0",
