@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { FormGroup, FormControl, Validators, ValidationErrors } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormControl, Validators, ValidationErrors } from '@angular/forms';
 
 
 import {AuthService} from '../auth.service';
@@ -17,18 +17,18 @@ export class RegisterComponent implements OnInit {
   ngOnInit() {
   }
 
-  passwordsMatchValidator(control: FormControl): ValidationErrors {
+  passwordsMatchValidator(control: UntypedFormControl): ValidationErrors {
     let password = control.root.get('password');
     return password && control.value !== password.value ? {
       passwordMatch: true
     }: null;
   }
 
-  userForm = new FormGroup({
-    fullname: new FormControl('', [Validators.required]),
-    email: new FormControl('', [Validators.required, Validators.email]),
-    password: new FormControl('', [Validators.required]),
-    repeatPassword: new FormControl('', [Validators.required, this.passwordsMatchValidator])
+  userForm = new UntypedFormGroup({
+    fullname: new UntypedFormControl('', [Validators.required]),
+    email: new UntypedFormControl('', [Validators.required, Validators.email]),
+    password: new UntypedFormControl('', [Validators.required]),
+    repeatPassword: new UntypedFormControl('', [Validators.required, this.passwordsMatchValidator])
   })
 
   get fullname(): any { return this.userForm.get('fullname'); }
